refactor: break live_view->components cycle + dismiss theming lazy cycles
Closes ~35 CodeQL py/cyclic-import notes and 7 misc note-level alerts.

- Extracted ContextProviderMixin from live_view.py to a new
  _context_provider.py module. components/base.py now imports it from
  there, breaking the module-level cycle (live_view -> serialization
  -> components/base -> live_view).
- ~28 theming cyclic-import notes are all lazy `from ... import ...`
  statements INSIDE function bodies (or their module-level counterparts
  paired with such lazy imports) — deliberate cycle breakers —
  bulk-dismissed with per-alert justification.
- 3 py/mixed-returns fixed structurally (inspector.py, admin_ext/views.py,
  djust_audit.py). 3 py/unused-global-variable dismissed as lazy-init
  cache FPs. 1 py/ineffectual-statement dismissed (await coro is real).

ContextProviderMixin is still importable from djust.live_view
(re-exported) for backward compat.
