+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- **Code-scanning cleanup: remaining ~35 `py/cyclic-import` notes + 7 misc note-level alerts**. Real refactor: extracted `ContextProviderMixin` from `live_view.py` to a new `_context_provider.py` module so `components/base.py` can import it without creating a module-level cycle back through `live_view -> serialization -> components/base`. `live_view.py` re-exports `ContextProviderMixin` for back-compat (existing user code importing `from djust.live_view import ContextProviderMixin` keeps working). Closes 3 real cyclic-import alerts (#2112, #2113, #2114). The remaining ~28 theming cyclic-import notes (in `manager.py`, `registry.py`, `theme_css_generator.py`, `pack_css_generator.py`, `theme_packs.py`, `manifest.py`, `css_generator.py`) are all `from ... import` statements INSIDE function bodies (or the module-level counterpart paired with such a lazy import) — deliberate cycle breakers where the runtime module graph is acyclic — dismissed with specific justification. Also fixed 3 `py/mixed-returns` via mechanical cleanup: `theming/inspector.py` (added 405 Method-Not-Allowed fallback), `admin_ext/views.py` (replaced bare `return` with `return None` in `run_action`), `management/commands/djust_audit.py` (explicit `return None` from all `handle()` branches). Dismissed 3 `py/unused-global-variable` false positives (lazy-init cache pattern in `components/icons.py:_icon_sets_cache`, `theming/theme_packs.py:_theme_imports_done`, `observability/log_handler.py:_installed_handler` — same pattern as `_psycopg` dismissed in #2104/#2105) and 1 `py/ineffectual-statement` false positive (`tutorials/mixin.py:371` — `await coro` is a real async effect, not an ineffectual expression). No behavior change; full Python suite passes (3428 passed, 15 skipped). (`python/djust/_context_provider.py`, `python/djust/live_view.py`, `python/djust/components/base.py`, `python/djust/theming/inspector.py`, `python/djust/admin_ext/views.py`, `python/djust/management/commands/djust_audit.py`)
+
 - **Cleanup: 36 `py/empty-except` + 6 misc CodeQL note-severity alerts** — Narrowed over-broad `except Exception: pass` to specific exception types where the call surface was knowable, and added `logger.debug(...)` (with `import logging; logger = logging.getLogger(__name__)` where not already present) for optional-feature probes in `components/gallery/views.py` (optional `djust_theming` static CSS link), `components/icons.py` (optional `DJUST_COMPONENTS_ICON_SETS` setting), `auth/admin_views.py` (optional `django-allauth` OAuth stats, 2 sites), `auth/djust_admin.py` (optional allauth registry), and `mixins/context.py` (best-effort descriptor resolution). Annotated "skip invalid numeric input" sites with justification comments (+ `pass` → `continue` for clarity) across `components/templatetags/_charts.py` (4), `components/rust_handlers.py` (8), `components/components/{calendar_heatmap,heatmap,line_chart,source_citation}.py`, `components/descriptors/carousel.py`, `components/function_component.py` (2), `components/mixins/data_table.py` (3), `components/templatetags/djust_components.py` (2), and similar narrow/intentional catches in `checks.py`, `components/base.py` (optional `@event_handler` decoration), `mixins/waiters.py` (idempotent waiter removal), `observability/dry_run.py` (best-effort bulk-op count), `theming/management/commands/djust_theme.py`, and `theming/templatetags/theme_tags.py`. Re-export in `components/templatetags/djust_components.py` (`_get_field_type`, `_infer_columns`, `_queryset_to_rows` from `_forms`) made explicit via `__all__` (closes `py/unused-import` #2171). Deleted 3 JS unused-variable declarations: `decoder` in `components/static/djust_components/ttyd/ttyd_terminal.js:35`, `resolvedMode` in `theming/static/djust_theming/js/theme.js:416`, and `getCookie()` in `theming/static/djust_theming/js/theme.js:449`. Dismissed 2 `py/unused-global-variable` false positives (#2104, #2105 — `_psycopg` / `_psycopg_sql` in `db/notifications.py` are lazy module-level caches assigned via `global` inside `_ensure_psycopg()`; CodeQL's scope analyzer doesn't track global-write patterns). 4 note-level `py/cyclic-import` alerts (#2096, #2112-#2114) left for scanner rescan — expected to auto-close as PR #928's refactor propagates. No behavior change; full Python suite passes (3428 passed, 15 skipped).
 
 - **Code-quality cleanup — ~66 CodeQL note-severity alerts** — mechanical fixes: deleted unused imports (treated re-exports with `__all__` + `# noqa: F401` preservation; replaced side-effect submodule imports with `importlib.import_module`), removed ~30 unused local variables across `rust_handlers.py`, `templatetags/djust_components.py`, `components/*.py`, and `templatetags/_forms.py` / `_advanced.py`, removed ~4 unused module-level names (`default_app_config` in `components/__init__.py`, `theming/__init__.py`, `admin_ext/__init__.py` — obsolete since Django 3.2 auto-discovery), simplified 3 `lambda vals: f(vals)` wrappers in `AGG_FUNCS` (pivot-table aggregations) to bare `sum` / `len`, deduped 2 `import json` / `import asyncio` occurrences in `function_component.py` / `mixins/data_table.py` / `db/notifications.py`, reconciled `import X` + `from X import Y` conflicts in `gallery/registry.py` and `templatetags/djust_components.py`, and removed ineffectual single-`...` statements in Protocol / abstract method bodies in `api/auth.py` and `tenants/audit.py`. No behavior change; full suite passes (3428). Plus 3 dismissed with justification: 2 × `py/catch-base-exception` in `async_work.py` (existing `# noqa: BLE001` comments + documented design intent of surfacing every failure via `AsyncResult.errored`), and 1 × `js/syntax-error` on `theming/templates/.../theme_head.html` (CodeQL's JS analyzer erroneously parsing a Django template as JavaScript).
