+++ python/djust/live_view.py
@@ -11,6 +11,7 @@ from typing import Any, Callable, Dict, List, Optional, Union
 
 from django.views import View
 
+from ._context_provider import ContextProviderMixin  # noqa: F401  # re-exported for back-compat
 from .serialization import DjangoJSONEncoder  # noqa: F401
 from .session_utils import (  # noqa: F401
     DEFAULT_SESSION_TTL,
@@ -71,47 +72,6 @@ __all__ = [
     "Stream",
     "extract_template_variables",
 ]
-
-
-class ContextProviderMixin:
-    """Component/view context sharing — React Context API equivalent (v0.5.1).
-
-    Mixed into both :class:`LiveView` and :class:`LiveComponent` so provider
-    and consumer roles work at any depth of the render tree. Providers are
-    stored per-instance in ``_djust_context_providers`` and the consumer walks
-    the ``_djust_context_parent`` chain (set by
-    ``LiveComponent.set_parent``).
-    """
-
-    def provide_context(self, key: str, value: Any) -> None:
-        """Expose a value to descendant components under ``key``.
-
-        Descendants read via :meth:`consume_context`. Scoped to the current
-        render tree; reset at the start of each render via
-        :meth:`clear_context_providers`.
-        """
-        providers = self.__dict__.setdefault("_djust_context_providers", {})
-        providers[key] = value
-
-    def consume_context(self, key: str, default: Any = None) -> Any:
-        """Return the value provided under ``key`` by this node or any ancestor.
-
-        Walks ``_djust_context_parent`` upward; returns ``default`` if no
-        provider is found.
-        """
-        node = self
-        while node is not None:
-            providers = getattr(node, "_djust_context_providers", None)
-            if providers and key in providers:
-                return providers[key]
-            node = getattr(node, "_djust_context_parent", None)
-        return default
-
-    def clear_context_providers(self) -> None:
-        """Reset all context providers — called at the start of each render."""
-        providers = self.__dict__.get("_djust_context_providers")
-        if providers:
-            providers.clear()
 
 
 class LiveView(
